Revert z-stack feature list, update progress bar
The z-stack seems to be working as I anticipate. I performed a 2 channel, 2 timepoint, 2 position acquisition. It acquired as follows:
for t:
    for p:
        for c:
             for z:

Perhaps this is because I added the timepoint information to the z-stack feature itself (see pre_data_func)? It may have been terminating early previously, but when I tested it without timepoints there it also worked.